pocx: fix narrowing in CompressedHeader nHeight initializer
CompressedHeader::nHeight is uint64_t, but the constructor brace-initializes
it from CBlockHeader::nHeight (int): nHeight{header.nHeight}. Braced
initialization forbids the narrowing int -> uint64_t conversion, so clang
rejects it under -Wc++11-narrowing / -Werror, breaking bitcoin_node and
cascading into every CI job (tidy, sanitizers, cross builds).

Add an explicit static_cast<uint64_t>, matching the implicit conversion the
v30 constructor performed via assignment (nHeight = header.nHeight). Height
is always non-negative, so the cast is value-preserving.

+++ src/headerssync.h
@@ -53,7 +53,7 @@ struct CompressedHeader {
           nTime{header.nTime},
 #ifdef ENABLE_POCX
           generationSignature{header.generationSignature},
-          nHeight{header.nHeight},
+          nHeight{static_cast<uint64_t>(header.nHeight)},
           nBaseTarget{header.nBaseTarget},
           pocxProof{header.pocxProof},
           vchPubKey{header.vchPubKey},
